+++ src/amqcsl/clients/bundles/pages.py
@@ -9,7 +9,7 @@ from attrs import Attribute, define, field, frozen
 from attrs.validators import deep_iterable, gt, instance_of
 
 from amqcsl.exceptions import QueryError
-from amqcsl.objects._db_types import CSLArtistSample, CSLGroup, CSLList, CSLSongSample, CSLTrack
+from amqcsl.objects import CSLArtistSample, CSLGroup, CSLList, CSLSongSample, CSLTrack
 from amqcsl.objects._json_types import JSONType, QueryArtist, QuerySong, QueryTrack
 
 from .core import LazyBundle, LazyMultiVendor, LazySingleVendor, LazyVendor, RichReprRtn, httpxClient
